+++ cmp/grid/GridModel.js
@@ -176,6 +176,7 @@ export class GridModel {
      * @param {boolean} [c.stripeRows] - true (default) to use alternating backgrounds for rows.
      * @param {boolean} [c.cellBorders] - true to render cell borders.
      * @param {boolean} [c.showCellFocus] - true to highlight the focused cell with a border.
+     * @param {boolean} [c.hideHeaders] - true to suppress display of the grid's header row.
      * @param {boolean} [c.enableColumnPinning] - true to allow the user to manually pin / unpin
      *      columns via UI affordances.
      * @param {boolean} [c.enableColChooser] - true to setup support for column chooser UI and
@@ -238,6 +239,7 @@ export class GridModel {
         cellBorders = false,
         stripeRows = true,
         showCellFocus = false,
+        hideHeaders = false,
         compact,
 
         enableColumnPinning = true,
@@ -307,7 +309,8 @@ export class GridModel {
             rowBorders,
             stripeRows,
             cellBorders,
-            showCellFocus
+            showCellFocus,
+            hideHeaders
         });
 
         this.colChooserModel = enableColChooser ? this.createChooserModel() : null;
@@ -453,6 +456,9 @@ export class GridModel {
 
     get showCellFocus() { return this.agGridModel.showCellFocus }
     setShowCellFocus(showCellFocus) { this.agGridModel.setShowCellFocus(showCellFocus) }
+
+    get hideHeaders() { return this.agGridModel.hideHeaders }
+    setHideHeaders(hideHeaders) { this.agGridModel.setHideHeaders(hideHeaders) }
 
     /**
      * Apply full-width row-level grouping to the grid for the given column ID(s).